cpufreq/arm-bl-cpufreq: Expose the frequency table to clients
This patch registers the frequency table for each CPU so that the
table and transition statistics are visible through sysfs, and to
governor code.

Since the frequency table is static and never deallocated while the
driver is loaded, there is no call to cpufreq_frequency_table_put_
attr().  A .exit method which calls this function would be needed
if the frequency table ever becomes dynamic.

Signed-off-by: Vishal Bhoj <vishal.bhoj@linaro.org>
Signed-off-by: Dave Martin <dave.martin@linaro.org>