ARM: perf: allocation of cpu_pmu at init time
cpu_pmu is a global pointer which is address of the one of arm_pmu
structure appropriately after the probing the PMU on the running CPU
or based on DTB bindings.
This patch removes all the static structure definations for every CPU
PMU type(e.g. armv7, armv6, armv6mp, xscale1, xscale2).
The structure is now allocted and populated during the initialization.
cpu_pmu global pointer is still not being removed.

Signed-off-by: Sudeep KarkadaNagesha <Sudeep.KarkadaNagesha@arm.com>